requirements.txt
pillow==10.4.0
pip-chill==1.0.3
pyqt5==5.15.11
